pyproject.toml
[project]
name = "llm-bias-eval"
version = "0.1.0"
description = "Alignment vs. Latent Bias in Large Language Models — bachelor thesis (UCPH)"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [{ name = "Sigurd Kornelius Havstein", email = "siggikornelius@gmail.com" }]

dependencies = [
    # Core ML
    "torch>=2.1",
    "transformers>=4.51",
    "accelerate>=0.30",
    "safetensors>=0.4",
    "sentencepiece>=0.2",
    "protobuf>=4.25",

    # Benchmarks
    "lm-eval>=0.4.0",

    # Probing
    "scikit-learn>=1.4",
    "numpy>=1.26,<2.2",
    "scipy>=1.12",

    # Data + analysis
    "datasets>=2.18",
    "huggingface-hub>=0.23",
    "pandas>=2.2",
    "pyarrow>=15",
    "matplotlib>=3.8",
    "seaborn>=0.13",

    # Tracking
    "wandb>=0.17",

    # Utilities
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "jsonlines>=4.0",
    "python-dotenv>=1.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "pre-commit>=3.7",
    "ipykernel>=6.29",
    "jupyter>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/biaseval"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "scripts", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",   # pycodestyle / pyflakes
    "I",             # isort
    "B",             # bugbear
    "UP",            # pyupgrade
    "N",             # pep8-naming
    "RUF",           # ruff-specific
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "N803",   # arg lowercase (X, y are sklearn convention)
    "N806",   # variable lowercase (X, y are sklearn convention)
    "RUF001", # ambiguous unicode in strings (× / − / – are intentional in log messages)
    "RUF002", # ambiguous unicode in docstrings
    "RUF003", # ambiguous unicode in comments
    "E702",   # multiple statements on one line — used in compact plotting code
]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "slow: tests that download models or hit GPU",
    "integration: end-to-end smoke tests",
]

[dependency-groups]
dev = []
